Moves Tests GetManagedLabels,GetManagedAnnotations
This moves the tests for GetManagedLabels and GetManagedAnnotations to
their respective util packages, as that's where they're now implemented.
