+++ util/annotations/helpers_test.go
@@ -17,12 +17,16 @@ limitations under the License.
 package annotations
 
 import (
+	"fmt"
+	"regexp"
 	"testing"
 
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
+	"sigs.k8s.io/cluster-api/util/test/builder"
 )
 
 func TestAddAnnotations(t *testing.T) {
@@ -284,5 +288,141 @@ func TestHasTruthyAnnotationValue(t *testing.T) {
 				g.Expect(ret).To(BeFalse())
 			}
 		})
+	}
+}
+
+func TestGetManagedAnnotations(t *testing.T) {
+	machineObj := newFakeMachine("default", "test-cluster")
+	ms := builder.MachineSet("default", "ms").Build()
+	ref := metav1.NewControllerRef(ms, ms.GroupVersionKind())
+
+	defaultAnnotations := map[string]string{
+		clusterv1.ClusterNameAnnotation:      machineObj.Spec.ClusterName,
+		clusterv1.ClusterNamespaceAnnotation: machineObj.GetNamespace(),
+		clusterv1.MachineAnnotation:          machineObj.Name,
+	}
+
+	additionalAnnotations := map[string]string{
+		"foo":                                "bar",
+		"bar":                                "baz",
+		"example.test/node.cluster.x-k8s.io": "not-managed",
+		"gpu-node.cluster.x-k8s.io":          "not-managed",
+		"example.test/node-restriction.kubernetes.io": "not-managed",
+		"gpu-node-restriction.kubernetes.io":          "not-managed",
+		"wrong.test.foo.example.com":                  "",
+	}
+
+	exampleRegex := regexp.MustCompile(`foo`)
+	defaultAndRegexAnnotations := map[string]string{}
+	for k, v := range defaultAnnotations {
+		defaultAndRegexAnnotations[k] = v
+	}
+	defaultAndRegexAnnotations["foo"] = "bar"
+	defaultAndRegexAnnotations["wrong.test.foo.example.com"] = ""
+
+	ownerRefAnnotations := map[string]string{
+		clusterv1.OwnerKindAnnotation: ms.Kind,
+		clusterv1.OwnerNameAnnotation: ms.Name,
+	}
+	defaultAndOwnerRefAnnotations := map[string]string{}
+	for k, v := range defaultAnnotations {
+		defaultAndOwnerRefAnnotations[k] = v
+	}
+	for k, v := range ownerRefAnnotations {
+		defaultAndOwnerRefAnnotations[k] = v
+	}
+
+	allAnnotations := map[string]string{}
+	for k, v := range defaultAnnotations {
+		allAnnotations[k] = v
+	}
+	for k, v := range additionalAnnotations {
+		allAnnotations[k] = v
+	}
+	for k, v := range ownerRefAnnotations {
+		allAnnotations[k] = v
+	}
+
+	tests := []struct {
+		name                             string
+		additionalSyncMachineAnnotations []*regexp.Regexp
+		allAnnotations                   map[string]string
+		managedAnnotations               map[string]string
+		owned                            bool
+	}{
+		{
+			name:                             "always sync default annotations",
+			additionalSyncMachineAnnotations: nil,
+			allAnnotations:                   allAnnotations,
+			managedAnnotations:               defaultAnnotations,
+		},
+		{
+			name: "sync additional defined labels",
+			additionalSyncMachineAnnotations: []*regexp.Regexp{
+				exampleRegex,
+			},
+			allAnnotations:     allAnnotations,
+			managedAnnotations: defaultAndRegexAnnotations,
+		},
+		{
+			name: "sync all labels",
+			additionalSyncMachineAnnotations: []*regexp.Regexp{
+				regexp.MustCompile(`.*`),
+			},
+			allAnnotations:     allAnnotations,
+			managedAnnotations: allAnnotations,
+		},
+		{
+			name:               "sync owner annotations",
+			allAnnotations:     allAnnotations,
+			managedAnnotations: defaultAndOwnerRefAnnotations,
+			owned:              true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testMachine := machineObj.DeepCopy()
+
+			if tt.owned {
+				testMachine.SetOwnerReferences([]metav1.OwnerReference{*ref})
+			}
+
+			testMachine.SetAnnotations(tt.allAnnotations)
+
+			g := NewWithT(t)
+			got := GetManagedAnnotations(testMachine, tt.additionalSyncMachineAnnotations...)
+			g.Expect(got).To(BeEquivalentTo(tt.managedAnnotations))
+		})
+	}
+}
+
+func newFakeMachineSpec(namespace, clusterName string) clusterv1.MachineSpec {
+	return clusterv1.MachineSpec{
+		ClusterName: clusterName,
+		Bootstrap: clusterv1.Bootstrap{
+			ConfigRef: &corev1.ObjectReference{
+				APIVersion: "bootstrap.cluster.x-k8s.io/v1alpha3",
+				Kind:       "KubeadmConfigTemplate",
+				Name:       fmt.Sprintf("%s-md-0", clusterName),
+				Namespace:  namespace,
+			},
+		},
+		InfrastructureRef: corev1.ObjectReference{
+			APIVersion: "infrastructure.cluster.x-k8s.io/v1alpha3",
+			Kind:       "FakeMachineTemplate",
+			Name:       fmt.Sprintf("%s-md-0", clusterName),
+			Namespace:  namespace,
+		},
+	}
+}
+
+func newFakeMachine(namespace, clusterName string) *clusterv1.Machine {
+	return &clusterv1.Machine{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      fmt.Sprintf("ma-annotationtest"),
+			Namespace: namespace,
+		},
+		Spec: newFakeMachineSpec(namespace, clusterName),
 	}
 }
