+++ util/labels/helpers_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package labels
 
 import (
+	"regexp"
 	"testing"
 
 	. "github.com/onsi/gomega"
@@ -137,6 +138,86 @@ func TestIsMachinePoolOwned(t *testing.T) {
 
 			result := IsMachinePoolOwned(tt.object)
 			g.Expect(result).To(Equal(tt.expected))
+		})
+	}
+}
+
+func TestGetManagedLabels(t *testing.T) {
+	defaultLabels := map[string]string{
+		clusterv1.NodeRoleLabelPrefix + "/anyRole": "",
+
+		clusterv1.ManagedNodeLabelDomain:                                  "",
+		"custom-prefix." + clusterv1.ManagedNodeLabelDomain:               "",
+		clusterv1.ManagedNodeLabelDomain + "/anything":                    "",
+		"custom-prefix." + clusterv1.ManagedNodeLabelDomain + "/anything": "",
+
+		clusterv1.NodeRestrictionLabelDomain:                                  "",
+		"custom-prefix." + clusterv1.NodeRestrictionLabelDomain:               "",
+		clusterv1.NodeRestrictionLabelDomain + "/anything":                    "",
+		"custom-prefix." + clusterv1.NodeRestrictionLabelDomain + "/anything": "",
+	}
+
+	additionalLabels := map[string]string{
+		"foo":                               "bar",
+		"bar":                               "baz",
+		"company.xyz/node.cluster.x-k8s.io": "not-managed",
+		"gpu-node.cluster.x-k8s.io":         "not-managed",
+		"company.xyz/node-restriction.kubernetes.io": "not-managed",
+		"gpu-node-restriction.kubernetes.io":         "not-managed",
+		"wrong.test.foo.com":                         "",
+	}
+
+	exampleRegex := regexp.MustCompile(`foo`)
+	defaultAndRegexLabels := map[string]string{}
+	for k, v := range defaultLabels {
+		defaultAndRegexLabels[k] = v
+	}
+	defaultAndRegexLabels["foo"] = "bar"
+	defaultAndRegexLabels["wrong.test.foo.com"] = ""
+
+	allLabels := map[string]string{}
+	for k, v := range defaultLabels {
+		allLabels[k] = v
+	}
+	for k, v := range additionalLabels {
+		allLabels[k] = v
+	}
+
+	tests := []struct {
+		name                        string
+		additionalSyncMachineLabels []*regexp.Regexp
+		allLabels                   map[string]string
+		managedLabels               map[string]string
+	}{
+		{
+			name:                        "always sync default labels",
+			additionalSyncMachineLabels: nil,
+			allLabels:                   allLabels,
+			managedLabels:               defaultLabels,
+		},
+		{
+			name: "sync additional defined labels",
+			additionalSyncMachineLabels: []*regexp.Regexp{
+				exampleRegex,
+			},
+			allLabels:     allLabels,
+			managedLabels: defaultAndRegexLabels,
+		},
+		{
+			name: "sync all labels",
+			additionalSyncMachineLabels: []*regexp.Regexp{
+				regexp.MustCompile(`.*`),
+			},
+			allLabels:     allLabels,
+			managedLabels: allLabels,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			g := NewWithT(t)
+			got := GetManagedLabels(tt.allLabels, tt.additionalSyncMachineLabels...)
+			g.Expect(got).To(BeEquivalentTo(tt.managedLabels))
 		})
 	}
 }
